// packages/components/src/ButtonBase/enums.ts
export enum ButtonColors {
  Error = 'error',
  Success = 'success',
  Warning = 'warning',
  Primary = 'primary',
}

export enum ButtonVariants {
  Contained = 'contained',
  Light = 'light',
  Text = 'text',
  Link = 'link',
}

export enum ButtonSizes {
  Small = 'small',
  Medium = 'medium',
  Large = 'large',
}

export enum ButtonStates {
  Default = 'default',
  Active = 'active',
  Focus = 'focus',
  Hover = 'hover',
}